hotplug: optimize put_cpu_down() and sampling rates
The code might seem to be more complex, but it should be
more efficient then the previous one. Each case gets handled
very surely.

Also changing the behaviour of the sampling rates. The tunable
is now a multiplier of HZ to ensure consitency. This fixes
also a bug where the sampling rate could go wild in some cases.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>