hotplug: reduce overhead for avg load
Before this patch when we did our load calculations, we could
end up in some sort of faked data. We would calculate the load
for all cpus 2^2^2 times (in the best case with only two cpus
online). When all cpus are online we would end up in some sort
of strange data.

Also introduce a new tunable min_online_time which sets
the minimum amount of time for a cpu to stay online.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>