hotplug: add a tunable for all_cpus_threshold
Previously adding a tunable for this parameter would end up in
some sort of deadlock as soon as the device enters deepsleep.
This is now avoided by allocating the right amount of memory
for the hotplug data.
Since the hotplug driver is enabled by default and shouldn't be
replaced by userspace we don't need to bother about freeing up the
memory.

This commit also contains a big cleanup for various things.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>