hotplug: reduce the amount of calculations
In a standard scenario we increase (or decrease) the load-counter
for each cpu two times in one sample rate. Clearly thats not the
point, because for ony sampling rate there should be only one
decision made for a cpu (increase or decrease).

Also fixing a bug with the average load;

       load  load-history
  cpu0:  100         100
  cpu1:   50           0
  cpu2:  100         100
  cpu3: (off)        100
  ------------------------
  avg(real)    :    83,3
  avg(previous):    87,5

When the load drops in cpu0/1 another core would get offlined,
but we won't clear the history for this core. By fixing this we
should end up with a small battery life increase for light usage.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>