hotplug: add delayed workqueue for suspend
Some users reported to have some sort of choppy audio
while listening to music when the screen is off. Its likely
caused by only having a single core active.

To fix this, we start a second workqueue which bypasses
the "normal" hotplug logic to react really quick on load
spikes and onlines a second core.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>